+++ app/client-app/src/app/pages/Wordlist/index.jsx
@@ -1,6 +1,7 @@
 import { useEffect, useState } from "react";
 import { Box } from "@chakra-ui/react";
 import { WordCardList } from "./components/WordCardList";
+import { excludeBuiltinWords, includeBuiltinWords } from "api/wordlist";
 import LightHeading from "components/core/LightHeading";
 import adjectives from "services/adjectives";
 import animals from "services/animals";
@@ -8,10 +9,12 @@ import { fetchWordList } from "api/wordlist";
 import { Page } from "components/core";
 import { toDictionary } from "services/data-structures";
 import { getFilteredWordList } from "./components/helpers";
+import { useWordlistSortingAndFiltering } from "./components/useWordlistSortingAndFiltering";
 
 const Wordlist = () => {
   const [wordList, setWordList] = useState(null);
   const [cards, setCards] = useState([]);
+  const { sorting, onSort, outputList } = useWordlistSortingAndFiltering(cards);
 
   useEffect(() => {
     const getWordList = async () => {
@@ -42,6 +45,24 @@ const Wordlist = () => {
     }
   }, [wordList]);
 
+  const toggleExclude = async (word, type, isExcludedBuiltin) => {
+    const id = wordList.id;
+
+    if (isExcludedBuiltin) {
+      await includeBuiltinWords(id, type, word);
+    } else {
+      await excludeBuiltinWords(id, type, word);
+    }
+
+    setCards((prevCards) =>
+      prevCards.map((card) =>
+        card.word === word
+          ? { ...card, isExcludedBuiltin: !isExcludedBuiltin }
+          : card
+      )
+    );
+  };
+
   return (
     <Page layout="default">
       <Box p={2}>
@@ -49,7 +70,15 @@ const Wordlist = () => {
           My Wordlist
         </LightHeading>
 
-        {cards.length > 0 && <WordCardList cards={cards} />}
+        {cards.length > 0 && (
+          <WordCardList
+            cards={cards}
+            sorting={sorting}
+            onSort={onSort}
+            toggleExclude={toggleExclude}
+            outputList={outputList}
+          />
+        )}
       </Box>
     </Page>
   );
